v1_complete_logout_reply_userright_pagenation_like

+++ app/Http/Controllers/CategoryController.php
@@ -10,37 +10,58 @@ class CategoryController extends Controller
     //
     public function index()
     {
-        $categorys = Category::orderby('title','asc')->get();
-        return view('category.index')
-        ->with('categories',$categorys);
+        if(isset(auth()->user()->id)){
+            $categorys = Category::orderby('title','asc')->get();
+
+            return view('category.index')
+            ->with('categories',$categorys);
+        } else {
+            return redirect('/');
+        }
     }
     public function view($id)
     {
-        $category = Category::find($id);
-        return view('category.view')
-        ->with('category',$category);
+        if(isset(auth()->user()->id)){
+            $category = Category::find($id);
+            return view('category.view')
+            ->with('category',$category);
+        } else {
+            return redirect('/');
+        }
     }
     public function store(Request $request)
     {
-        $category=new Category;
-        $category->title = $request->title;
-        $category->save();
+        if(isset(auth()->user()->id)){
+            $category=new Category;
+            $category->title = $request->title;
+            $category->save();
 
-        return redirect('/category');
+            return redirect('/category');
+        } else {
+            return redirect('/');
+        }
     }
     public function update(Request $request,$id)
     {
         $category = Category::find($id);
-        $category->title = $request->title;
-        $category->save();
+        if(isset(auth()->user()->id)){
+            $category->title = $request->title;
+            $category->save();
 
-        return redirect('/category');
+            return redirect('/category');
+        } else {
+            return redirect('/');
+        }
     }
     public function delete($id)
     {
-        $category = Category::find($id);
-        $category->delete();
-        return redirect('/category');
+        if(isset(auth()->user()->id)){
+            $category = Category::find($id);
+            $category->delete();
+            return redirect('/category');
+        } else {
+            return redirect('/');
+        }
 
     }
 }
